code cleanup

+++ batch_layer_rename/index.js
@@ -1,3 +1,5 @@
+const { app } = window.require('photoshop');
+
 // Helper function to execute photoshop as modal.
 function executePhotoshopModal(callback, commandName) {
   return window.require('photoshop').core.executeAsModal(callback, { commandName });
@@ -16,66 +18,42 @@ function getInputValue(selector) {
 function renameLayerNames() {
   const newName = getInputValue('#rename_layer_text');
   if (!newName) return;
-
-  return executePhotoshopModal(() => {
-    const app = window.require('photoshop').app;
-
-    app.activeDocument.activeLayers.forEach((layer) => {
-      layer.name = newName;
-    });
+  executePhotoshopModal(() => {
+    app.activeDocument.activeLayers.forEach((layer) => (layer.name = newName));
   }, 'Rename layers');
 }
 
 function addPrefix() {
   const prefixText = getInputValue('#add_prefix_text');
   if (!prefixText) return;
-
-  return executePhotoshopModal(() => {
-    const app = window.require('photoshop').app;
-
-    app.activeDocument.activeLayers.forEach((layer) => {
-      layer.name = `${prefixText}_${layer.name}`;
-    });
+  executePhotoshopModal(() => {
+    app.activeDocument.activeLayers.forEach((layer) => (layer.name = `${prefixText}_${layer.name}`));
   }, 'Add Prefix');
 }
 
 function addSuffix() {
   const suffixText = getInputValue('#add_suffix_text');
   if (!suffixText) return;
-
-  return executePhotoshopModal(() => {
-    const app = window.require('photoshop').app;
-    app.activeDocument.activeLayers.forEach((layer) => {
-      layer.name = `${layer.name}_${suffixText}`;
-    });
+  executePhotoshopModal(() => {
+    app.activeDocument.activeLayers.forEach((layer) => (layer.name = `${layer.name}_${suffixText}`));
   }, 'Add Suffix');
 }
 
 // DELETE EMPTY LAYERS FUNCTIONALITY //
 async function isEmpty(layer) {
   if (layer.layers) {
-    // Recursively check if all sub-layers are empty
-    return (await Promise.all(layer.layers.map(async (subLayer) => await isEmpty(subLayer)))).every(
-      (isEmpty) => isEmpty
-    );
-  } else {
-    const bounds = layer.bounds;
-    const boundsWidth = bounds.right - bounds.left;
-    const boundsHeight = bounds.bottom - bounds.top;
-    const isEmpty = boundsWidth === 0 || boundsHeight === 0;
-    return isEmpty;
+    const results = await Promise.all(layer.layers.map(isEmpty));
+    return results.every((result) => result);
   }
+  const { left, right, top, bottom } = layer.bounds;
+  return right - left === 0 || bottom - top === 0;
 }
 
+// Recursively delete empty layers
 async function deleteEmptyLayers(layers) {
   for (const layer of layers) {
-    if (layer.layers) {
-      // Recursively delete empty sub-layers
-      await deleteEmptyLayers(layer.layers);
-    }
-    if (await isEmpty(layer)) {
-      layer.delete();
-    }
+    if (layer.layers) await deleteEmptyLayers(layer.layers);
+    if (await isEmpty(layer)) layer.delete();
   }
 }
 
@@ -84,9 +62,7 @@ async function deleteEmpty() {
   if (!confirmed) return;
 
   await executePhotoshopModal(async () => {
-    const app = window.require('photoshop').app;
-    const layers = app.activeDocument.layers;
-    await deleteEmptyLayers(layers);
+    await deleteEmptyLayers(app.activeDocument.layers);
   }, 'Delete Layers');
 }
